Scale highlight/selection halo with 'shrink on zoom' things

+++ src/MapEditor/Renderer/2D/MapRenderer2D.h
@@ -165,7 +165,7 @@ private:
 	// Misc.
 	void    updateThingPaths(const vector<MapThing*>& things);
 	void    setupThingBuffer(gl::ThingBuffer2D& buffer, const game::ThingType& tt) const;
-	double  scaledRadius(int radius) const;
+	double  scaledRadius(double radius) const;
 	ColRGBA lineColour(const MapLine* line, bool ignore_filter = false) const;
 };
 } // namespace slade::mapeditor
